Move warp histogram functionality to TelemetryRecorder Use newer CMake requirement & set CUDA standard to C++17
