+++ Apps/InfiniTAM_cli/CMakeLists.txt
@@ -12,7 +12,6 @@ SET(targetname InfiniTAM_cli)
 # Specify the libraries to use #
 ################################
 
-include(${PROJECT_SOURCE_DIR}/cmake/UseCUDA.cmake)
 include(${PROJECT_SOURCE_DIR}/cmake/UseLibRoyale.cmake)
 include(${PROJECT_SOURCE_DIR}/cmake/UseOpenGL.cmake)
 include(${PROJECT_SOURCE_DIR}/cmake/UseOpenMP.cmake)
