+++ ITMLib/Engines/Telemetry/CPU/TelemetryRecorder_CPU_VoxelBlockHash.cpp
@@ -1,6 +1,6 @@
 //  ================================================================
-//  Created by Gregory Kramida on 8/27/19.
-//  Copyright (c) 2019 Gregory Kramida
+//  Created by Gregory Kramida on 7/13/18.
+//  Copyright (c) 2018-2000 Gregory Kramida
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
 //  You may obtain a copy of the License at
@@ -13,9 +13,12 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 //  ================================================================
-#include "VolumeSequenceRecorder.tpp"
+
+#include "../../../GlobalTemplateDefines.h"
+#include "../TelemetryRecorder.tpp"
+#include "../../Traversal/CPU/VolumeTraversal_CPU_VoxelBlockHash.h"
 
 namespace ITMLib {
 template
-class VolumeSequenceRecorder<TSDFVoxel, WarpVoxel, VoxelBlockHash>;
+class TelemetryRecorder<TSDFVoxel, WarpVoxel, VoxelBlockHash, MEMORYDEVICE_CPU>;
 } // namespace ITMLib
