+++ Tests/CMakeLists.txt
@@ -5,7 +5,6 @@
 option(BUILD_TESTS "Build automated tests with the project (Boost-test required)." OFF)
 
 macro(load_use_includes)
-    include(${PROJECT_SOURCE_DIR}/cmake/UseCUDA.cmake)
     include(${PROJECT_SOURCE_DIR}/cmake/UseLibRoyale.cmake)
     include(${PROJECT_SOURCE_DIR}/cmake/UseOpenGL.cmake)
     include(${PROJECT_SOURCE_DIR}/cmake/UseOpenMP.cmake)
